Refactor to prevent circular dependency
