+++ src/app/containers/MediaPlayer/index.jsx
@@ -7,7 +7,7 @@ import {
 import Caption from '../Caption';
 import Metadata from './Metadata';
 import embedUrl from './helpers/embedUrl';
-import getPlaceholderSrc from './helpers/placeholder';
+import getPlaceholderSrc from '#lib/utilities/srcSet/placeholder';
 import { getPlaceholderSrcSet } from '#lib/utilities/srcSet';
 import filterForBlockType from '#lib/utilities/blockHandlers';
 import useToggle from '../Toggle/useToggle';
